# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Interv API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/userInWorkspace.delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userInWorkspace"
                ],
                "summary": "Delete User From Workspace",
                "operationId": "DeleteUserFromWorkspace",
                "parameters": [
                    {
                        "description": "DeleteUserFromWorkspaceBody",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DeleteUserFromWorkspaceBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response-UserInWorkspace"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
        "/userInWorkspace.get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userInWorkspace"
                ],
                "summary": "Get user In Workspace",
                "operationId": "GetUserInWorkspace",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response-array_UserInWorkspace"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace.create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Create new workspace",
                "operationId": "CreateWorkspace",
                "parameters": [
                    {
                        "description": "CreateWorkspaceBody",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateWorkspaceBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response-WorkspaceDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
        "/workspace.delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Delete workspace By Id",
                "operationId": "DeleteWorkspaceById",
                "parameters": [
                    {
                        "description": "DeleteWorkspaceBody",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DeleteWorkspaceBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
        "/workspace.get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Get workspace",
                "operationId": "GetWorkspace",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response-WorkspaceData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
        "/workspace.getByPortal": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Get List of workspace",
                "operationId": "GetPortalWorkspace",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response-array_WorkspaceDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CreateWorkspaceBody": {
            "type": "object",
            "required": [
                "isCoding",
                "isVideo",
                "startDate",
                "stopDate",
                "title"
            ],
            "properties": {
                "isCoding": {
                    "type": "boolean"
                },
                "isVideo": {
                    "type": "boolean"
                },
                "startDate": {
                    "type": "string"
                },
                "stopDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "DeleteUserFromWorkspaceBody": {
            "type": "object",
            "required": [
                "userId",
                "workspaceId"
            ],
            "properties": {
                "userId": {
                    "type": "integer"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
        "DeleteWorkspaceBody": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "IndividualUser": {
            "type": "object",
            "required": [
                "id",
                "userData",
                "userInWorkspace"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "userData": {
                    "$ref": "#/definitions/User"
                },
                "userInWorkspace": {
                    "$ref": "#/definitions/UserInWorkspace"
                }
            }
        },
...
        "User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "UserInWorkspace": {
            "type": "object",
            "required": [
                "id",
                "isInterest",
                "status",
                "userId",
                "workspaceId"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isInterest": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
...
        "WorkspaceData": {
            "type": "object",
            "required": [
                "individualUser",
                "workspaceDetail"
            ],
            "properties": {
                "individualUser": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IndividualUser"
                    }
                },
                "workspaceDetail": {
                    "$ref": "#/definitions/WorkspaceDetail"
                }
            }
        },
        "WorkspaceDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isCoding": {
                    "type": "boolean"
                },
                "isVideo": {
                    "type": "boolean"
                },
                "memberNum": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "stopDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "handlers.Response-UserInWorkspace": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/UserInWorkspace"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "handlers.Response-WorkspaceData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/WorkspaceData"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handlers.Response-WorkspaceDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/WorkspaceDetail"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "handlers.Response-array_UserInWorkspace": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/UserInWorkspace"
                    }
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handlers.Response-array_WorkspaceDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/WorkspaceDetail"
                    }
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "handlers.Response-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}